Multiple cleanUps and different improvements for performance. ReadMe Rewrite
