+++ src/app/components/ui/SearchMediaCard.tsx
@@ -1,0 +1,52 @@
+import { useSearchParams } from "next/navigation";
+import Image from "next/image";
+import { IDetailedMedia } from "../../interfaces";
+import { BorderContainer } from "../../utils/borderContainer";
+
+export function SearchMediaCard({
+  selectedmedia,
+  media,
+  updateMedia,
+}: {
+  selectedmedia: IDetailedMedia;
+  media: IDetailedMedia;
+  updateMedia: (stream_name: string, tmdb_id: number) => void;
+}) {
+  const stream_name = useSearchParams().get("stream_name") || "";
+
+  if (media) {
+    return (
+      <BorderContainer className={`w-full h-64`}>
+        <div className={`flex gap-4 w-full h-full rounded-md`}>
+          <Image
+            src={"https://image.tmdb.org/t/p/original" + media.poster}
+            width={300}
+            height={200}
+            alt=""
+            className="h-full w-1/6 rounded-md"
+          />
+          <div className="flex flex-col justify-between w-full h-full">
+            <h1 className="text-2xl font-bold">{media.name}</h1>
+            <>
+              <p className="text-sm">Original Name: {media.original_name}</p>
+              <p className="text-sm">ID: {media.tmdb_id}</p>
+              <p className="text-sm">First Air Date: {media.first_air_date}</p>
+            </>
+            <button
+              className={
+                selectedmedia.tmdb_id === media.tmdb_id
+                  ? "hidden"
+                  : "customButton"
+              }
+              onClick={() => updateMedia(stream_name, media.tmdb_id)}
+            >
+              Select This Medium
+            </button>
+          </div>
+        </div>
+      </BorderContainer>
+    );
+  } else {
+    return <div className={`w-full h-36 bg-gray-900/50 rounded-md`} />;
+  }
+}
