+++ src/app/layout.tsx
@@ -4,6 +4,7 @@ import "./globals.css";
 import ClientLayout from "./components/clientlayout";
 import { getConfiguration } from "./actions/configurationProvider";
 import SetupPage from "./components/setupPage";
+import Image from "next/image";
 
 const geistSans = localFont({
   src: "./fonts/GeistVF.woff",
@@ -20,7 +21,7 @@ export async function generateMetadata(): Promise<Metadata> {
   const appConfig = await getConfiguration();
   return {
     title: `${appConfig.appName}`,
-    icons: "favicon.ico",
+    icons: { icon: "icon.png", apple: "apple-icon.png" },
     description: `${appConfig.appName} version: ${appConfig.appVersion}`,
   };
 }
@@ -37,25 +38,24 @@ export default async function RootLayout({
       id="html"
       lang="en"
       className={`max-w-screen max-h-screen bg-gray-900 no-text-cursor`}
-      style={
-        appConfig.background_image
-          ? {
-              backgroundImage: "url('/assets/wallpaper')",
-              backgroundPosition: "center",
-              backgroundSize: "cover",
-              backgroundAttachment: "fixed",
-              backgroundColor: "rgba(17, 24, 39, 0.75)",
-              backgroundBlendMode: "overlay",
-              backdropFilter: "blur(4px)",
-              zIndex: -1,
-            }
-          : {}
-      }
     >
       <body
         id="body"
         className={`flex h-screen w-full text-gray-300  ${geistSans.variable} ${geistMono.variable} antialiased`}
       >
+        {appConfig.background_image && (
+          <div className="fixed inset-0 -z-50">
+            <Image
+              src="/assets/wallpaper"
+              alt=""
+              fill={true}
+              className="object-cover bg-no-repeat"
+              priority={true}
+              unoptimized={true}
+            />
+            <div className="absolute inset-0 bg-gray-950/75" />
+          </div>
+        )}
         {appConfig.configured ? (
           <ClientLayout appConfig={appConfig}>{children}</ClientLayout>
         ) : (
