+++ src/app/series-overview/page.tsx
@@ -2,10 +2,11 @@
 
 import { useEffect, useState } from "react";
 import { IMedia } from "../interfaces";
-import { BackdropMediaCard, PosterMediaCard } from "../utils/mediaCards";
 import { SearchBar } from "../utils/searchBar";
 import { ScrollContainer } from "../components/scrollToRef";
 import { useAppConfigContext } from "../utils/appConfigContext";
+import { BackdropMediaCard } from "../components/ui/BackdropMediaCard";
+import { PosterMediaCard } from "../components/ui/PosterMediaCard";
 
 export default function SeriesOverview() {
   const appConfig = useAppConfigContext();
@@ -83,7 +84,7 @@ export default function SeriesOverview() {
       <p className="my-4 mb-8 mx-[25%] border border-gray-400" />
       <div id="seriesContainer" className="grid grid-cols-5 gap-4">
         {mediaList.map((media, index) => (
-          <PosterMediaCard key={index} media={media} className="h-96" />
+          <PosterMediaCard key={index} media={media} />
         ))}
       </div>
     </ScrollContainer>
