+++ src/app/settings/components/SettingsContainer.tsx
@@ -9,7 +9,7 @@ export function SettingsContainer({
 }) {
   return (
     <div
-      className={`${className} flex flex-col w-full h-[36rem] mb-4 bg-gray-300/25 backdrop-blur-sm p-4 rounded-md`}
+      className={`${className} flex flex-col w-full h-[36rem] mb-4 p-4 rounded-md`}
     >
       <h1 className="secondHeaddline">{title}</h1>
       {children}
